+++ components/card/card.module.css
@@ -1,0 +1,3 @@
+.work svg {
+    vertical-align: top;
+}
