+++ public/global.css
@@ -38,7 +38,3 @@
     transform: none;
     transition: transform 225ms cubic-bezier(0.4, 0, 0.2, 1) 0ms;
 }
-
-.work-card-link svg {
-    vertical-align: top;
-}
